libft_0

+++ CommonCore/0_libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: falves-b <falves-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 18:46:22 by falves-b          #+#    #+#             */
-/*   Updated: 2022/11/09 17:36:38 by falves-b         ###   ########.fr       */
+/*   Updated: 2022/11/21 00:27:53 by falves-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*tmp;
