+++ CommonCore/0_libft/ft_memchr.c
@@ -6,14 +6,14 @@
 /*   By: falves-b <falves-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 18:49:37 by falves-b          #+#    #+#             */
-/*   Updated: 2022/11/09 18:02:03 by falves-b         ###   ########.fr       */
+/*   Updated: 2022/11/21 00:38:20 by falves-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <string.h>
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*tmp;
